refactor(apps): update transparency log components
