+++ apps/moderation/src/components/transparency-log.tsx
@@ -5,17 +5,10 @@ import DS from '@akashaproject/design-system';
 import { moderationRequest } from '@akashaproject/ui-awf-hooks';
 
 import { ICount } from './content-list';
+import Banner from './transparency-log/banner';
+import DetailCard, { ILogItem } from './transparency-log/detail-card';
 
-const {
-  Box,
-  Text,
-  Icon,
-  Spinner,
-  SwitchCard,
-  TransparencyLogBanner,
-  TransparencyLogMiniCard,
-  TransparencyLogDetailCard,
-} = DS;
+const { Box, Text, Icon, Spinner, SwitchCard, TransparencyLogMiniCard } = DS;
 
 export interface ITransparencyLogProps {
   user: string | null;
@@ -24,24 +17,6 @@ export interface ITransparencyLogProps {
   navigateToUrl: (url: string) => void;
 }
 
-interface ILogItem {
-  contentID: string;
-  contentType: string;
-  moderatedDate: Date;
-  moderator: {
-    ethAddress: string;
-    name: string;
-    userName: string;
-    avatar: string;
-  };
-  delisted: false;
-  reasons: string[];
-  reports: number;
-  explanation: string;
-}
-
-const BASE_SOCIAL_URL = '/social-app';
-const BASE_PROFILE_URL = '/profile';
 const DEFAULT_LIMIT = 10;
 
 const TransparencyLog: React.FC<ITransparencyLogProps> = props => {
@@ -71,7 +46,7 @@ const TransparencyLog: React.FC<ITransparencyLogProps> = props => {
       if (requesting) return;
       if (listItemObserver.current) listItemObserver.current.disconnect();
       listItemObserver.current = new IntersectionObserver(entries => {
-        if (entries[0].isIntersecting && nextIndex !== null) {
+        if (entries[0].isIntersecting && nextIndex) {
           // fetch more entries using nextIndex
           fetchModerationLog(nextIndex);
         }
@@ -127,16 +102,6 @@ const TransparencyLog: React.FC<ITransparencyLogProps> = props => {
     setSelected(null);
   };
 
-  const handleClickViewItem = (contentType: string, contentID: string) => () => {
-    if (contentType === 'post') {
-      navigateToUrl(`${BASE_SOCIAL_URL}/post/${contentID}`);
-    } else if (contentType === 'reply' || contentType === 'comment') {
-      navigateToUrl(`${BASE_SOCIAL_URL}/reply/${contentID}`);
-    } else if (contentType === 'account') {
-      navigateToUrl(`${BASE_PROFILE_URL}/${contentID}`);
-    }
-  };
-
   const handleClickAvatar = () => {
     /* TODO */
   };
@@ -151,63 +116,6 @@ const TransparencyLog: React.FC<ITransparencyLogProps> = props => {
   const buttonLabels = !isMobile
     ? [t('All'), t('Kept'), t('Delisted')]
     : [t('All'), t('Kept'), t('Delisted'), t('Stats')];
-
-  const RenderDetailCard = (selected: ILogItem) => {
-    return (
-      <TransparencyLogDetailCard
-        locale="en"
-        title={t(
-          `${selected.contentType.charAt(0).toUpperCase()}${selected.contentType.substring(1)} ${
-            selected.delisted ? 'delisted by' : 'kept by'
-          }`,
-        )}
-        content={t(`${selected.explanation}`)}
-        isDelisted={selected.delisted}
-        moderator={selected.moderator.name}
-        moderatedTimestamp={selected.moderatedDate.toString()}
-        moderatorAvatarUrl={`${ipfsGateway}/${selected.moderator.avatar}`}
-        moderatorEthAddress={selected.moderator.ethAddress}
-        reportedTimesLabel={t(
-          `Reported ${selected.reports > 1 ? `${selected.reports} times` : 'once'}`,
-        )}
-        viewItemLink={`${window.location.origin}${
-          selected.contentType === 'account'
-            ? `${BASE_PROFILE_URL}/${selected.contentID}`
-            : `${BASE_SOCIAL_URL}/${selected.contentType}/${selected.contentID}`
-        }`}
-        viewItemLabel={t(`View ${selected.contentType}`)}
-        reasonsLabel={t(`${selected.reasons.length > 1 ? 'reasons' : 'reason'}`)}
-        reasons={selected.reasons}
-        explanationLabel={t('explanation')}
-        contactModeratorsLabel={t('Contact the moderators')}
-        contactModeratorsLink="mailto:moderators@ethereum.world"
-        onClickArrowLeft={handleClickArrowLeft}
-        onClickViewItem={handleClickViewItem(selected.contentType, selected.contentID)}
-        onClickAvatar={handleClickAvatar}
-      />
-    );
-  };
-
-  const RenderBanner = () => {
-    return (
-      <TransparencyLogBanner
-        size="12.75rem"
-        assetName="moderation-history-illustration"
-        title={t('Moderation History')}
-        content={t(
-          'Here you will find the moderated posts, replies, and accounts of Ethereum World. We do not reveal any personal information of the author or submitter(s) to protect their privacy.',
-        )}
-        keptCount={count.kept}
-        keptCountLabel={t('kept')}
-        totalCountLabel={t('total')}
-        delistedCount={count.delisted}
-        delistedCountLabel={t('delisted')}
-        footerLabel={t('Visit our Code of Conduct to learn more about our moderation criteria')}
-        footerLinkLabel={t('Code of Conduct')}
-        footerLink="https://akasha.ethereum.world/legal/code-of-conduct"
-      />
-    );
-  };
 
   return (
     <Box>
@@ -235,35 +143,37 @@ const TransparencyLog: React.FC<ITransparencyLogProps> = props => {
             <Text>{t('No moderated items found. Please check again later.')}</Text>
           )}
           {logItems.length > 0 &&
-            (activeButton === 'Stats'
-              ? RenderBanner()
-              : logItems
-                  .filter(el =>
-                    activeButton === 'Kept'
-                      ? !el.delisted
-                      : activeButton === 'Delisted'
-                      ? el.delisted
-                      : el,
-                  )
-                  .map((el, idx) => (
-                    <TransparencyLogMiniCard
-                      key={idx}
-                      locale="en"
-                      title={t(
-                        `${el.contentType.charAt(0).toUpperCase()}${el.contentType.substring(1)} ${
-                          el.delisted ? 'Delisted' : 'Kept'
-                        }`,
-                      )}
-                      content={t(`${el.explanation}`)}
-                      isSelected={el.contentID === selected?.contentID}
-                      isDelisted={el.delisted}
-                      moderatedTimestamp={el.moderatedDate.toString()}
-                      moderatorAvatarUrl={`${ipfsGateway}/${el.moderator.avatar}`}
-                      moderatorEthAddress={el.moderator.ethAddress}
-                      onClickAvatar={handleClickAvatar}
-                      onClickCard={handleClickCard(el)}
-                    />
-                  )))}
+            (activeButton === 'Stats' ? (
+              <Banner count={count} />
+            ) : (
+              logItems
+                .filter(el =>
+                  activeButton === 'Kept'
+                    ? !el.delisted
+                    : activeButton === 'Delisted'
+                    ? el.delisted
+                    : el,
+                )
+                .map((el, idx) => (
+                  <TransparencyLogMiniCard
+                    key={idx}
+                    locale="en"
+                    title={t(
+                      `${el.contentType.charAt(0).toUpperCase()}${el.contentType.substring(1)} ${
+                        el.delisted ? 'Delisted' : 'Kept'
+                      }`,
+                    )}
+                    content={t(`${el.explanation}`)}
+                    isSelected={el.contentID === selected?.contentID}
+                    isDelisted={el.delisted}
+                    moderatedTimestamp={el.moderatedDate.toString()}
+                    moderatorAvatarUrl={`${ipfsGateway}/${el.moderator.avatar}`}
+                    moderatorEthAddress={el.moderator.ethAddress}
+                    onClickAvatar={handleClickAvatar}
+                    onClickCard={handleClickCard(el)}
+                  />
+                ))
+            ))}
           {/* fetch indicator for initial page load */}
           {requesting && logItems.length < 1 && (
             <Box pad="large">
@@ -282,11 +192,28 @@ const TransparencyLog: React.FC<ITransparencyLogProps> = props => {
         </Box>
         {isMobile && selected && (
           <Box width="100%" style={{ position: 'absolute', right: 0 }}>
-            {RenderDetailCard(selected)}
+            <DetailCard
+              selected={selected}
+              ipfsGateway={ipfsGateway}
+              handleClickAvatar={handleClickAvatar}
+              handleClickArrowLeft={handleClickArrowLeft}
+              navigateToUrl={navigateToUrl}
+            />
           </Box>
         )}
         {!isMobile && (
-          <Box width="60%">{selected ? RenderDetailCard(selected) : RenderBanner()}</Box>
+          <Box width="60%">
+            {selected && (
+              <DetailCard
+                selected={selected}
+                ipfsGateway={ipfsGateway}
+                handleClickAvatar={handleClickAvatar}
+                handleClickArrowLeft={handleClickArrowLeft}
+                navigateToUrl={navigateToUrl}
+              />
+            )}
+            {!selected && <Banner count={count} />}
+          </Box>
         )}
       </Box>
     </Box>
